+++ include/rf2o_laser_odometry/CLaserOdometry2D.h
@@ -34,15 +34,6 @@
 #include <Eigen/Geometry>
 #include <unsupported/Eigen/MatrixFunctions>
 
-#include <tf2/convert.h>
-#include <tf2/exceptions.h>
-#include <tf2_ros/transform_broadcaster.h>
-#include <tf2_ros/transform_listener.h>
-#include <tf2_ros/buffer.h>
-#include <tf2/impl/utils.h>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
-#include <tf2/utils.h>
-
 namespace rf2o {
 
 template <typename T>
