+++ launch/rf2o_laser_odometry.launch.py
@@ -16,7 +16,6 @@ def generate_launch_description():
 
     return LaunchDescription([
 
-            # if we don't use camera:
             Node(
                 package='rf2o_laser_odometry',
                 executable='rf2o_laser_odometry_node',
@@ -26,7 +25,7 @@ def generate_launch_description():
                     'laser_scan_topic' : '/scan',
                     'odom_topic' : '/odom_rf2o',
                     'publish_tf' : False,
-                    'base_frame_id' : 'base_link',
+                    'base_frame_id' : 'base_footprint',
                     'odom_frame_id' : 'odom',
                     'init_pose_from_topic' : '',
                     'freq' : 6.0}],
