+++ src/CLaserOdometry2DNode.cpp
@@ -15,102 +15,9 @@
 * Modifications: Jeremie Deray
 ******************************************************************************************** */
 
-#include "rf2o_laser_odometry/CLaserOdometry2D.h"
+#include "rf2o_laser_odometry/CLaserOdometry2DNode.h"
 
-
-namespace rf2o {
-
-class CLaserOdometry2DNode : public rclcpp::Node
-{
-public:
-  CLaserOdometry2DNode();
-  void process();
-  void publish();
-  bool setLaserPoseFromTf();
-
-  CLaserOdometry2D rf2o_ref;
-  bool publish_tf, new_scan_available;
-
-  double freq;
-
-  std::string         laser_scan_topic;
-  std::string         odom_topic;
-  std::string         base_frame_id;
-  std::string         odom_frame_id;
-  std::string         init_pose_from_topic;
-
-  sensor_msgs::msg::LaserScan      last_scan;
-  bool                        GT_pose_initialized;
-  std::shared_ptr<tf2_ros::Buffer> buffer_;
-  std::shared_ptr<tf2_ros::TransformListener> tf_listener_;  
-  std::unique_ptr<tf2_ros::TransformBroadcaster> odom_broadcaster;
-  nav_msgs::msg::Odometry     initial_robot_pose;
-
-  //Subscriptions & Publishers
-  rclcpp::Subscription<sensor_msgs::msg::LaserScan>::SharedPtr  laser_sub;
-  rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr  initPose_sub;
-  rclcpp::Publisher<nav_msgs::msg::Odometry>::SharedPtr  odom_pub;
-
-  bool scan_available();
-
-  //CallBacks
-  void LaserCallBack(const sensor_msgs::msg::LaserScan::SharedPtr new_scan);
-  void initPoseCallBack(const nav_msgs::msg::Odometry::SharedPtr new_initPose);
-};
-
-CLaserOdometry2DNode::CLaserOdometry2DNode(): Node("CLaserOdometry2DNode")
-{
-  RCLCPP_INFO(get_logger(), "Initializing RF2O node...");
-
-  //Read Parameters
-  //----------------
-  this->declare_parameter<std::string>("laser_scan_topic", "/laser_scan");
-  this->get_parameter("laser_scan_topic", laser_scan_topic);
-  this->declare_parameter<std::string>("odom_topic", "/odom_rf2o");
-  this->get_parameter("odom_topic", odom_topic);
-  this->declare_parameter<std::string>("base_frame_id", "/base_link");
-  this->get_parameter("base_frame_id", base_frame_id);
-  this->declare_parameter<std::string>("odom_frame_id", "/odom");
-  this->get_parameter("odom_frame_id", odom_frame_id);
-  this->declare_parameter<bool>("publish_tf", true);
-  this->get_parameter("publish_tf", publish_tf);
-  this->declare_parameter<std::string>("init_pose_from_topic", "/base_pose_ground_truth");
-  this->get_parameter("init_pose_from_topic", init_pose_from_topic);
-  this->declare_parameter<double>("freq", 10.0);
-  this->get_parameter("freq", freq);
-
-  //Publishers and Subscribers
-  //--------------------------
-  buffer_ = std::make_shared<tf2_ros::Buffer>(this->get_clock());
-  tf_listener_ = std::make_shared<tf2_ros::TransformListener>(*buffer_);
-  odom_broadcaster = std::make_unique<tf2_ros::TransformBroadcaster>(this);
-  odom_pub  = this->create_publisher<nav_msgs::msg::Odometry>(odom_topic, 5);
-  laser_sub = this->create_subscription<sensor_msgs::msg::LaserScan>(laser_scan_topic,rclcpp::QoS(rclcpp::KeepLast(1)).best_effort().durability_volatile(),
-      std::bind(&CLaserOdometry2DNode::LaserCallBack, this, std::placeholders::_1));
-  //init pose??
-  if (init_pose_from_topic != "")
-  {
-    initPose_sub = this->create_subscription<nav_msgs::msg::Odometry>(init_pose_from_topic,rclcpp::QoS(rclcpp::KeepLast(1)).best_effort().durability_volatile(),
-        std::bind(&CLaserOdometry2DNode::initPoseCallBack, this, std::placeholders::_1));
-    GT_pose_initialized  = false;
-  }
-  else
-  {
-    GT_pose_initialized = true;
-    initial_robot_pose.pose.pose.position.x = 0;
-    initial_robot_pose.pose.pose.position.y = 0;
-    initial_robot_pose.pose.pose.position.z = 0;
-    initial_robot_pose.pose.pose.orientation.w = 0;
-    initial_robot_pose.pose.pose.orientation.x = 0;
-    initial_robot_pose.pose.pose.orientation.y = 0;
-    initial_robot_pose.pose.pose.orientation.z = 0;
-  }
-
-
-  //Init variables
-  rf2o_ref.module_initialized = false;
-  rf2o_ref.first_laser_scan   = true;
-}
+using namespace rf2o;
 
 bool CLaserOdometry2DNode::setLaserPoseFromTf()
 {
